+++ bencher/__init__.py
@@ -21,6 +21,11 @@ from bencher.results.holoview_results.table_result import TableResult
 from bencher.results.holoview_results.tabular_spec import TabularSpec
 from bencher.results.holoview_results.tabulator_result import TabulatorResult
 from bencher.results.holoview_results.xy_curve_result import XYCurveResult, xy_curve
+from bencher.results.holoview_results.xy_hexbin_result import XYHexbinResult, xy_hexbin
+from bencher.results.holoview_results.xy_histogram_result import (
+    XYHistogramResult,
+    xy_histogram,
+)
 from bencher.results.holoview_results.xy_scatter_result import XYScatterResult, xy_scatter
 from bencher.results.volume_result import VolumeResult